Make Button block a functional component
