+++ src/hoc/index.js
@@ -1,7 +1,9 @@
 import withUniqueId from './withUniqueId';
 import withGridLegacyMigration from './withGridLegacyMigration';
+import withButtonLegacyMigration from './withButtonLegacyMigration';
 
 export {
 	withUniqueId,
 	withGridLegacyMigration,
+	withButtonLegacyMigration,
 };
